fix(iot-e2e): enable TransportClientIT on Android
